% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwfit.R
\name{AIC.gkwfit}
\alias{AIC.gkwfit}
\title{Calculate AIC for a gkwfit object}
\usage{
\method{AIC}{gkwfit}(object, ..., k = 2)
}
\arguments{
\item{object}{A gkwfit object}

\item{...}{Additional models}

\item{k}{Penalty per parameter}
}
\description{
Calculate AIC for a gkwfit object
}
