% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwreg.R
\name{AIC.gkwreg}
\alias{AIC.gkwreg}
\title{Akaike's Information Criterion for Generalized Kumaraswamy Regression Models}
\usage{
\method{AIC}{gkwreg}(object, ..., k = 2)
}
\arguments{
\item{object}{An object of class \code{"gkwreg"}, typically the result of a call
to \code{\link{gkwreg}}.}

\item{...}{Optionally, additional objects of class \code{"gkwreg"} for model
comparison.}

\item{k}{The penalty per parameter to be used; the default \code{k = 2} gives
the traditional AIC.}
}
\value{
If just one object is provided, returns a numeric AIC value. If multiple
objects are provided, returns a data frame with rows corresponding to the objects
and columns representing the number of parameters (df) and AIC values.
}
\description{
Calculates the Akaike Information Criterion (AIC) for one or more fitted
Generalized Kumaraswamy regression models. AIC is a measure of the relative
quality of statistical models for a given set of data, providing a means for
model selection.
}
\details{
The AIC is calculated as:

\deqn{AIC = -2 \times \log(L) + 2 \times k}

where \eqn{L} is the maximized likelihood function and \eqn{k} is the number of
estimated parameters in the model. The model with the lowest AIC is generally
preferred.

For small sample sizes, the corrected AIC (AICc) may be more appropriate.
Though not directly implemented here, it can be calculated as:

\deqn{AICc = AIC + \frac{2k(k+1)}{n-k-1}}

where \eqn{n} is the sample size.
}
\examples{
\dontrun{
# Fit two competing models
kw_reg1 <- gkwreg(y ~ x1 | x2, data = df, family = "kw")
kw_reg2 <- gkwreg(y ~ x1 + x3 | x2, data = df, family = "kw")

# Compare models using AIC
AIC(kw_reg1, kw_reg2)

# Calculate AIC with a different penalty
AIC(kw_reg1, k = 3)
}

}
\references{
Akaike, H. (1974). A new look at the statistical model identification.
\emph{IEEE Transactions on Automatic Control}, 19(6), 716-723.
}
\seealso{
\code{\link{gkwreg}}, \code{\link{logLik.gkwreg}}, \code{\link{BIC.gkwreg}}
}
