% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwreg.R
\name{AIC}
\alias{AIC}
\title{Register the S3 method}
\usage{
AIC(object, ...)
}
\arguments{
\item{object}{gkwreg class fit}

\item{...}{Additional arguments passed to or from other methods.}
}
\description{
Register the S3 method
}
