% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwfitall.R
\name{calculate_fit_metrics}
\alias{calculate_fit_metrics}
\title{Calculate additional fit metrics for all models}
\usage{
calculate_fit_metrics(fits, data)
}
\arguments{
\item{fits}{List of fitted gkwfit models}

\item{data}{Original data vector}
}
\value{
List with additional fit metrics
}
\description{
Calculate additional fit metrics for all models
}
\keyword{internal}
