% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculateCoxSnellResiduals}
\alias{calculateCoxSnellResiduals}
\title{Calculate Cox-Snell Residuals}
\usage{
calculateCoxSnellResiduals(y, alpha, beta, gamma, delta, lambda)
}
\arguments{
\item{y}{NumericVector of observations.}

\item{alpha}{NumericVector of alpha parameters.}

\item{beta}{NumericVector of beta parameters.}

\item{gamma}{NumericVector of gamma parameters.}

\item{delta}{NumericVector of delta parameters.}

\item{lambda}{NumericVector of lambda parameters.}
}
\value{
NumericVector of Cox-Snell residuals.
}
\description{
Computes Cox-Snell residuals defined as -log(1 - F(y)), where F is the cumulative distribution function.
}
