% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculatePearsonResiduals}
\alias{calculatePearsonResiduals}
\title{Calculate Pearson Residuals}
\usage{
calculatePearsonResiduals(y, fitted, alpha, beta, gamma, delta, lambda)
}
\arguments{
\item{y}{NumericVector of observations.}

\item{fitted}{NumericVector of fitted values (means).}

\item{alpha}{NumericVector of alpha parameters.}

\item{beta}{NumericVector of beta parameters.}

\item{gamma}{NumericVector of gamma parameters.}

\item{delta}{NumericVector of delta parameters.}

\item{lambda}{NumericVector of lambda parameters.}
}
\value{
NumericVector of Pearson residuals.
}
\description{
Computes the Pearson residuals based on the observed values, fitted means, and the approximate variance of the distribution.
}
