% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculateResponseResiduals}
\alias{calculateResponseResiduals}
\title{Calculate Response Residuals}
\usage{
calculateResponseResiduals(y, fitted)
}
\arguments{
\item{y}{NumericVector of observations.}

\item{fitted}{NumericVector of fitted values.}
}
\value{
NumericVector of response residuals.
}
\description{
Computes the raw response residuals as the difference between the observed and fitted values.
}
