% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwfitall.R
\name{create_comparison_table}
\alias{create_comparison_table}
\title{Create comparison table of fit statistics}
\usage{
create_comparison_table(fits)
}
\arguments{
\item{fits}{List of fitted gkwfit models}
}
\value{
Data frame with comparison statistics
}
\description{
Create comparison table of fit statistics
}
\keyword{internal}
