% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwfitall.R
\name{create_residual_plot}
\alias{create_residual_plot}
\title{Create residual plot for all fitted models}
\usage{
create_residual_plot(fits, data)
}
\arguments{
\item{fits}{List of fitted gkwfit models}

\item{data}{Original data vector}
}
\value{
A ggplot2 object with the residual plot
}
\description{
Create residual plot for all fitted models
}
\keyword{internal}
