% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dbkw}
\alias{dbkw}
\title{Density Function for Beta-Kumaraswamy Distribution}
\usage{
dbkw(x, alpha, beta, gamma, delta, log_prob = FALSE)
}
\arguments{
\item{x}{Vector of quantiles in (0,1).}

\item{alpha}{Shape parameter α > 0.}

\item{beta}{Shape parameter β > 0.}

\item{gamma}{Shape parameter γ > 0.}

\item{delta}{Shape parameter δ ≥ 0.}

\item{log_prob}{Logical; if TRUE, returns log(f(x)).}
}
\value{
A vector of length max(length(x), lengths of parameters) with computed densities
or log-densities. Values of x outside (0,1) yield 0 or -Inf (log-scale).
}
\description{
Computes the PDF of the Beta-Kumaraswamy distribution BKw(α, β, γ, δ) for x in (0, 1).
Optionally returns the log of the PDF.
}
\details{
The PDF is given by
\deqn{
  f(x) = \frac{\alpha \beta}{B(\gamma, \delta+1)} \; x^{\alpha - 1} \; \bigl(1 - x^\alpha\bigr)^{\beta(\delta+1) - 1}\;
         \bigl[1 - \bigl(1 - x^\alpha\bigr)^\beta\bigr]^{\gamma - 1}, \quad 0<x<1.
}

Note that BKw is a special case of the Generalized Kumaraswamy (GKw) with \eqn{\lambda=1}.
}
