% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dekw}
\alias{dekw}
\title{Density for Exponentiated Kumaraswamy Distribution}
\usage{
dekw(x, alpha, beta, lambda, log_prob = FALSE)
}
\arguments{
\item{x}{Vector of quantiles in (0,1).}

\item{alpha}{Shape parameter \eqn{\alpha > 0}.}

\item{beta}{Shape parameter \eqn{\beta > 0}.}

\item{lambda}{Shape parameter \eqn{\lambda > 0}.}

\item{log_prob}{Logical; if TRUE, returns log-density, else density.}
}
\value{
A vector of the same length as the broadcast of (x, alpha, beta, lambda).
}
\description{
Computes the PDF of EKw(α, β, λ), where \eqn{\alpha, β, λ > 0}.
}
\details{
The PDF is
\deqn{
  f(x) = \lambda\,\alpha\,\beta \; x^{\alpha-1} (1 - x^α)^{\beta-1} \;
         [1 - (1 - x^α)^\beta ]^{\lambda - 1}, \quad 0<x<1.
}
This follows from the GKw distribution with \eqn{\gamma=1,\;\delta=0}.
}
