% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwreg.R
\name{.calculate_diagnostic_measures}
\alias{.calculate_diagnostic_measures}
\title{Calculate diagnostic measures for gkwreg plots}
\usage{
.calculate_diagnostic_measures(
  y_obs,
  fitted_vals,
  resid_vec,
  model_matrices,
  model_params,
  param_vectors,
  idx,
  family,
  param_info
)
}
\arguments{
\item{y_obs}{Vector of observed values}

\item{fitted_vals}{Vector of fitted values}

\item{resid_vec}{Vector of residuals}

\item{model_matrices}{List of model matrices}

\item{model_params}{List of model parameters}

\item{param_vectors}{List of parameter vectors}

\item{idx}{Vector of observation indices}

\item{family}{Character string specifying distribution family}

\item{param_info}{Parameter information for the family}
}
\value{
A list of diagnostic measures
}
\description{
Calculate diagnostic measures for gkwreg plots
}
\keyword{internal}
