% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwgof.R
\name{.calculate_distance_tests}
\alias{.calculate_distance_tests}
\title{Calculate Distance-Based Test Statistics}
\usage{
.calculate_distance_tests(sorted_data, p_empirical, p_theoretical)
}
\arguments{
\item{sorted_data}{Sorted numeric vector of the data}

\item{p_empirical}{Numeric vector of empirical CDF values}

\item{p_theoretical}{Numeric vector of theoretical CDF values}
}
\value{
A list containing the calculated test statistics
}
\description{
Calculate Distance-Based Test Statistics
}
\keyword{internal}
