% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwgof.R
\name{.calculate_information_criteria}
\alias{.calculate_information_criteria}
\title{Calculate Information Criteria}
\usage{
.calculate_information_criteria(loglik, k, n)
}
\arguments{
\item{loglik}{Log-likelihood value}

\item{k}{Number of parameters}

\item{n}{Sample size}
}
\value{
A list containing various information criteria
}
\description{
Calculate Information Criteria
}
\keyword{internal}
