% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwreg.R
\name{.calculate_model_parameters}
\alias{.calculate_model_parameters}
\title{Calculate model parameters for the specified family}
\usage{
.calculate_model_parameters(model_matrices, model_params, family)
}
\arguments{
\item{model_matrices}{List of model matrices}

\item{model_params}{List of model parameters}

\item{family}{Character string specifying distribution family}
}
\value{
A matrix of calculated parameters
}
\description{
Calculate model parameters for the specified family
}
\keyword{internal}
