% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwgof.R
\name{.calculate_prediction_metrics}
\alias{.calculate_prediction_metrics}
\title{Calculate Prediction Accuracy Metrics}
\usage{
.calculate_prediction_metrics(p_empirical, p_theoretical)
}
\arguments{
\item{p_empirical}{Numeric vector of empirical CDF values}

\item{p_theoretical}{Numeric vector of theoretical CDF values}
}
\value{
A list containing prediction accuracy metrics
}
\description{
Calculate Prediction Accuracy Metrics
}
\keyword{internal}
