% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwfit.R
\name{.calculate_profiles}
\alias{.calculate_profiles}
\title{Calculate profile likelihoods}
\usage{
.calculate_profiles(result, data, family, fixed, fit, method, npoints, silent)
}
\arguments{
\item{result}{Fit result from TMB or Newton-Raphson.}

\item{data}{Numeric vector with values in the (0, 1) interval.}

\item{family}{Character string specifying the distribution family.}

\item{fixed}{List of parameters to be held fixed.}

\item{fit}{Estimation method: "tmb" or "nr".}

\item{method}{Optimization method (for TMB): "nlminb" or "optim".}

\item{npoints}{Number of points in profile.}

\item{silent}{Logical; if TRUE, suppresses messages.}
}
\value{
List of profile likelihoods.
}
\description{
Calculate profile likelihoods
}
\keyword{internal}
