% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwreg.R
\name{.calculate_residuals}
\alias{.calculate_residuals}
\title{Calculate residuals based on the specified type}
\usage{
.calculate_residuals(y_obs, fitted_vals, param_mat, type, family)
}
\arguments{
\item{y_obs}{Vector of observed values}

\item{fitted_vals}{Vector of fitted values}

\item{param_mat}{Matrix of calculated parameters}

\item{type}{Character string specifying residual type}

\item{family}{Character string specifying distribution family}
}
\value{
A vector of residuals
}
\description{
Calculate residuals based on the specified type
}
\keyword{internal}
