% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwgof.R
\name{.calculate_sample_moments}
\alias{.calculate_sample_moments}
\title{Calculate Sample Moments}
\usage{
.calculate_sample_moments(data)
}
\arguments{
\item{data}{Numeric vector of data}
}
\value{
A vector containing the first four sample moments
}
\description{
Calculate Sample Moments
}
\keyword{internal}
