% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwreg.R
\name{.calculate_sim_residuals}
\alias{.calculate_sim_residuals}
\title{Calculate residuals for simulated data}
\usage{
.calculate_sim_residuals(sim_y, param_mat, type, family)
}
\arguments{
\item{sim_y}{Vector of simulated observations}

\item{param_mat}{Matrix of calculated parameters}

\item{type}{Character string specifying residual type}

\item{family}{Character string specifying distribution family}
}
\value{
A vector of residuals
}
\description{
Calculate residuals for simulated data
}
\keyword{internal}
