% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwgof.R
\name{.calculate_theoretical_pdf}
\alias{.calculate_theoretical_pdf}
\title{Calculate Theoretical PDF Values for GKw Family Distributions}
\usage{
.calculate_theoretical_pdf(x, family, params)
}
\arguments{
\item{x}{Numeric vector of data points}

\item{family}{Character string specifying the distribution family}

\item{params}{Named numeric vector of distribution parameters}
}
\value{
Numeric vector of PDF values
}
\description{
Calculate Theoretical PDF Values for GKw Family Distributions
}
\keyword{internal}
