% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwgof.R
\name{.calculate_theoretical_quantiles}
\alias{.calculate_theoretical_quantiles}
\title{Calculate Theoretical Quantiles for GKw Family Distributions}
\usage{
.calculate_theoretical_quantiles(p, family, params)
}
\arguments{
\item{p}{Numeric vector of probabilities}

\item{family}{Character string specifying the distribution family}

\item{params}{Named numeric vector of distribution parameters}
}
\value{
Numeric vector of quantiles
}
\description{
Calculate Theoretical Quantiles for GKw Family Distributions
}
\keyword{internal}
