% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwreg-utils.R
\name{.convert_links_to_int}
\alias{.convert_links_to_int}
\title{Convert Link Function Names to TMB Integers}
\usage{
.convert_links_to_int(link_list)
}
\arguments{
\item{link_list}{List of link function names}
}
\value{
List of link function integers for TMB
}
\description{
Convert Link Function Names to TMB Integers
}
\keyword{internal}
