% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwgof.R
\name{.create_radar_plot}
\alias{.create_radar_plot}
\title{Create Radar Plot for Model Comparison}
\usage{
.create_radar_plot(gof_list, criteria)
}
\arguments{
\item{gof_list}{List of gkwgof objects}

\item{criteria}{Vector of criteria to include in the plot}
}
\value{
A radar plot
}
\description{
Create Radar Plot for Model Comparison
}
\keyword{internal}
