% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwgof.R
\name{.create_table_plot}
\alias{.create_table_plot}
\title{Create Table Plot for Model Comparison}
\usage{
.create_table_plot(data, title)
}
\arguments{
\item{data}{Data frame containing the comparison data}

\item{title}{Table title}
}
\value{
A ggplot2 object
}
\description{
Create Table Plot for Model Comparison
}
\keyword{internal}
