% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwreg.R
\name{.extract_model_matrices}
\alias{.extract_model_matrices}
\title{Extract model matrices from a gkwreg object with family-specific handling}
\usage{
.extract_model_matrices(x)
}
\arguments{
\item{x}{A fitted model object of class "gkwreg"}
}
\value{
A list of model matrices
}
\description{
Extract model matrices from a gkwreg object with family-specific handling
}
\keyword{internal}
