% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwreg.R
\name{.extract_model_params}
\alias{.extract_model_params}
\title{Extract model parameters from a gkwreg object with family-specific handling}
\usage{
.extract_model_params(x)
}
\arguments{
\item{x}{A fitted model object of class "gkwreg"}
}
\value{
A list of model parameters
}
\description{
Extract model parameters from a gkwreg object with family-specific handling
}
\keyword{internal}
