% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwreg.R
\name{.extract_parameter_vectors}
\alias{.extract_parameter_vectors}
\title{Extract parameter vectors from parameter matrix}
\usage{
.extract_parameter_vectors(param_mat)
}
\arguments{
\item{param_mat}{Matrix of calculated parameters}
}
\value{
A list of parameter vectors
}
\description{
Extract parameter vectors from parameter matrix
}
\keyword{internal}
