% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwfit.R
\name{.family_to_code}
\alias{.family_to_code}
\title{Convert family string to numeric code for TMB}
\usage{
.family_to_code(family)
}
\arguments{
\item{family}{Character string specifying the family.}
}
\value{
Integer code for TMB. 0="gkw", 1="bkw", 2="kkw", 3="ekw", 4="mc", 5="kw", 6="beta"
}
\description{
Convert family string to numeric code for TMB
}
\keyword{internal}
