% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwreg.R
\name{.format_coefficient_names}
\alias{.format_coefficient_names}
\title{Format Coefficient Names Based on Family and Model Matrices}
\usage{
.format_coefficient_names(param_names, model_data, param_positions)
}
\arguments{
\item{param_names}{Names of parameters for the family}

\item{model_data}{Model data list including matrices}

\item{param_positions}{Parameter position mapping for the family}
}
\value{
Vector of formatted coefficient names
}
\description{
Format Coefficient Names Based on Family and Model Matrices
}
\keyword{internal}
