% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwgof.R
\name{.generate_additional_plots}
\alias{.generate_additional_plots}
\title{Generate Additional Diagnostic Plots Beyond Those in gkwfit}
\usage{
.generate_additional_plots(
  object,
  data,
  p_empirical,
  p_theoretical,
  theme = ggplot2::theme_classic(),
  ...
)
}
\arguments{
\item{object}{Object of class "gkwfit"}

\item{data}{Numeric vector of data}

\item{p_empirical}{Numeric vector of empirical CDF values}

\item{p_theoretical}{Numeric vector of theoretical CDF values}

\item{theme}{A ggplot theme, defaults to theme_classic()}

\item{...}{Additional arguments to be passed to plotting functions}
}
\value{
A list of ggplot2 objects for additional diagnostic plots
}
\description{
Generate Additional Diagnostic Plots Beyond Those in gkwfit
}
\keyword{internal}
