% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwgof.R
\name{.generate_random_samples}
\alias{.generate_random_samples}
\title{Generate Random Samples from GKw Family Distributions}
\usage{
.generate_random_samples(n, family, params)
}
\arguments{
\item{n}{Integer number of samples to generate}

\item{family}{Character string specifying the distribution family}

\item{params}{Named numeric vector of distribution parameters}
}
\value{
Numeric vector of random samples
}
\description{
Generate Random Samples from GKw Family Distributions
}
\keyword{internal}
