% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwfit.R
\name{.get_default_fixed}
\alias{.get_default_fixed}
\title{Get default fixed parameters for a family}
\usage{
.get_default_fixed(family)
}
\arguments{
\item{family}{Character string specifying the family.}
}
\value{
Named list of fixed parameters.
}
\description{
Get default fixed parameters for a family
}
\keyword{internal}
