% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwfit.R
\name{.get_default_start}
\alias{.get_default_start}
\title{Get default start values for a family}
\usage{
.get_default_start(family)
}
\arguments{
\item{family}{Character string specifying the family.}
}
\value{
Named list of starting values.
}
\description{
Get default start values for a family
}
\keyword{internal}
