% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwfit.R
\name{.get_family_fixed_defaults}
\alias{.get_family_fixed_defaults}
\title{Get default fixed parameters for each GKw family}
\usage{
.get_family_fixed_defaults(family)
}
\arguments{
\item{family}{Character string, the family name}
}
\value{
Named list of parameters that are fixed by default for this family
}
\description{
Get default fixed parameters for each GKw family
}
\keyword{internal}
