% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwreg-compile-tmb.R
\name{.is_dll_functional}
\alias{.is_dll_functional}
\title{Check if DLL is Loaded and Functional}
\usage{
.is_dll_functional(dll_name, verbose = FALSE)
}
\arguments{
\item{dll_name}{Base name of DLL}

\item{verbose}{Logical}
}
\value{
Logical
}
\description{
Check if DLL is Loaded and Functional
}
\keyword{internal}
