% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwreg-compile-tmb.R
\name{.load_dll}
\alias{.load_dll}
\title{Load TMB DLL}
\usage{
.load_dll(dll_path, dll_name, verbose = FALSE)
}
\arguments{
\item{dll_path}{Full path to DLL file}

\item{dll_name}{Base name of DLL}

\item{verbose}{Logical}
}
\value{
Logical, TRUE if successful
}
\description{
Loads a TMB DLL and registers it with TMB's symbol table.
CRITICAL: Never calls dyn.unload() to avoid finalization errors.
}
\keyword{internal}
