% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwreg.R
\name{.map_gkwreg_to_tmb_param}
\alias{.map_gkwreg_to_tmb_param}
\title{Map gkwreg parameter index to TMB parameter index}
\usage{
.map_gkwreg_to_tmb_param(object, param_idx)
}
\arguments{
\item{object}{A fitted model object of class "gkwreg"}

\item{param_idx}{Index of the parameter in the gkwreg coefficients vector}
}
\value{
The corresponding index in the TMB parameter vector, or NA if mapping fails
}
\description{
Map gkwreg parameter index to TMB parameter index
}
\keyword{internal}
