% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwreg.R
\name{.plot_base_r_cooks_distance}
\alias{.plot_base_r_cooks_distance}
\title{Plot Cook's distance (base R)}
\usage{
.plot_base_r_cooks_distance(diag_data, p_main, ...)
}
\description{
Plot Cook's distance (base R)
}
\keyword{internal}
