% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwreg.R
\name{.plot_base_r_half_normal}
\alias{.plot_base_r_half_normal}
\title{Plot half-normal plot (base R)}
\usage{
.plot_base_r_half_normal(diag_data, p_main, type, ...)
}
\description{
Plot half-normal plot (base R)
}
\keyword{internal}
