% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwreg.R
\name{.plot_base_r_leverage_vs_fitted}
\alias{.plot_base_r_leverage_vs_fitted}
\title{Plot leverage vs. fitted (base R)}
\usage{
.plot_base_r_leverage_vs_fitted(diag_data, p_main, ...)
}
\description{
Plot leverage vs. fitted (base R)
}
\keyword{internal}
