% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwreg.R
\name{.plot_base_r_predicted_vs_observed}
\alias{.plot_base_r_predicted_vs_observed}
\title{Plot predicted vs. observed (base R)}
\usage{
.plot_base_r_predicted_vs_observed(diag_data, p_main, ...)
}
\description{
Plot predicted vs. observed (base R)
}
\keyword{internal}
