% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwreg.R
\name{.plot_base_r_residuals_vs_linpred}
\alias{.plot_base_r_residuals_vs_linpred}
\title{Plot residuals vs. linear predictor (base R)}
\usage{
.plot_base_r_residuals_vs_linpred(diag_data, p_main, type, ...)
}
\description{
Plot residuals vs. linear predictor (base R)
}
\keyword{internal}
