% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwreg.R
\name{.plot_ggplot_cooks_distance}
\alias{.plot_ggplot_cooks_distance}
\title{Plot Cook's distance (ggplot2)}
\usage{
.plot_ggplot_cooks_distance(diag_data, p_main, sub.caption, theme_fn)
}
\description{
Plot Cook's distance (ggplot2)
}
\keyword{internal}
