% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwreg.R
\name{.plot_ggplot_half_normal}
\alias{.plot_ggplot_half_normal}
\title{Plot half-normal plot (ggplot2)}
\usage{
.plot_ggplot_half_normal(diag_data, p_main, sub.caption, type, theme_fn)
}
\arguments{
\item{diag_data}{Diagnostic data list}

\item{p_main}{Plot title}

\item{sub.caption}{Plot subtitle}

\item{type}{Residual type}

\item{theme_fn}{ggplot2 theme function}
}
\value{
A ggplot object
}
\description{
Plot half-normal plot (ggplot2)
}
\keyword{internal}
