% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwreg.R
\name{.plot_ggplot_leverage_vs_fitted}
\alias{.plot_ggplot_leverage_vs_fitted}
\title{Plot leverage vs. fitted (ggplot2)}
\usage{
.plot_ggplot_leverage_vs_fitted(diag_data, p_main, sub.caption, theme_fn)
}
\description{
Plot leverage vs. fitted (ggplot2)
}
\keyword{internal}
