% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwreg.R
\name{.plot_ggplot_predicted_vs_observed}
\alias{.plot_ggplot_predicted_vs_observed}
\title{Plot predicted vs. observed (ggplot2)}
\usage{
.plot_ggplot_predicted_vs_observed(diag_data, p_main, sub.caption, theme_fn)
}
\description{
Plot predicted vs. observed (ggplot2)
}
\keyword{internal}
