% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwreg.R
\name{.plot_ggplot_residuals_vs_index}
\alias{.plot_ggplot_residuals_vs_index}
\title{Plot residuals vs. index (ggplot2)}
\usage{
.plot_ggplot_residuals_vs_index(diag_data, p_main, sub.caption, type, theme_fn)
}
\description{
Plot residuals vs. index (ggplot2)
}
\keyword{internal}
