% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwreg.R
\name{.plot_ggplot_residuals_vs_linpred}
\alias{.plot_ggplot_residuals_vs_linpred}
\title{Plot residuals vs. linear predictor (ggplot2)}
\usage{
.plot_ggplot_residuals_vs_linpred(
  diag_data,
  p_main,
  sub.caption,
  type,
  theme_fn
)
}
\description{
Plot residuals vs. linear predictor (ggplot2)
}
\keyword{internal}
