% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwreg-core.R
\name{.prepare_tmb_data}
\alias{.prepare_tmb_data}
\title{Prepare TMB Data for GKw Regression}
\usage{
.prepare_tmb_data(
  model_data,
  family,
  param_names,
  fixed,
  link_ints,
  link_scale_list,
  y,
  param_positions
)
}
\arguments{
\item{model_data}{List of model data.}

\item{family}{Family name.}

\item{param_names}{Names of parameters.}

\item{fixed}{List of fixed parameters and coefficients.}

\item{link_ints}{List of link function integers.}

\item{link_scale_list}{List of link scale values.}

\item{y}{Response variable.}

\item{param_positions}{Parameter position mapping for the family.}
}
\value{
A list with TMB data.
}
\description{
Prepare TMB Data for GKw Regression
}
\keyword{internal}
