% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwgof.R
\name{.print_gof_summary}
\alias{.print_gof_summary}
\title{Print Formatted Summary of Goodness-of-Fit Statistics}
\usage{
.print_gof_summary(results, verbose)
}
\arguments{
\item{results}{List of results from gkwgof function}

\item{verbose}{Logical; if TRUE, provides additional details}
}
\value{
None (called for its side effect of printing to console)
}
\description{
Print Formatted Summary of Goodness-of-Fit Statistics
}
\keyword{internal}
