% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwreg.R
\name{.process_fixed}
\alias{.process_fixed}
\title{Process Fixed Parameters for GKw Regression}
\usage{
.process_fixed(fixed, param_names, fixed_params)
}
\arguments{
\item{fixed}{List of fixed parameters or coefficients.}

\item{param_names}{Names of the parameters for the specified family.}

\item{fixed_params}{List of fixed parameters from the family definition.}
}
\value{
A list of processed fixed parameters and coefficients.
}
\description{
Process Fixed Parameters for GKw Regression
}
\keyword{internal}
