% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwreg-utils.R
\name{.process_formula_parts}
\alias{.process_formula_parts}
\title{Process Formula Parts from a Formula Object}
\usage{
.process_formula_parts(formula_obj, param_names, fixed_params, data)
}
\arguments{
\item{formula_obj}{Formula object created with the Formula package.}

\item{param_names}{Names of the parameters for the specified family.}

\item{fixed_params}{List of fixed parameters.}

\item{data}{Data frame containing the variables.}
}
\value{
A list of formula objects for each parameter.
}
\description{
Process Formula Parts from a Formula Object
}
\keyword{internal}
