% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwreg-utils.R
\name{.process_link}
\alias{.process_link}
\title{Process Link Functions for GKw Regression}
\usage{
.process_link(link, param_names, fixed_params)
}
\arguments{
\item{link}{A character string or list of character strings specifying link functions.}

\item{param_names}{Names of the parameters for the specified family.}

\item{fixed_params}{List of fixed parameters.}
}
\value{
A list of link functions.
}
\description{
Process Link Functions for GKw Regression
}
\keyword{internal}
