% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwreg.R
\name{.process_link_scale}
\alias{.process_link_scale}
\title{Process Link Scales for GKw Regression}
\usage{
.process_link_scale(link_scale, link_list, param_names, fixed_params)
}
\arguments{
\item{link_scale}{A numeric value or list specifying scales for link functions.}

\item{link_list}{List of link functions for each parameter.}

\item{param_names}{Names of the parameters for the specified family.}

\item{fixed_params}{List of fixed parameters.}
}
\value{
A list of link scales.
}
\description{
Process Link Scales for GKw Regression
}
\keyword{internal}
