% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwreg.R
\name{.sample_model_data}
\alias{.sample_model_data}
\title{Sample model data for large datasets}
\usage{
.sample_model_data(n, sample_size, y_obs, fitted_vals, model_matrices)
}
\arguments{
\item{n}{Total number of observations}

\item{sample_size}{Target sample size}

\item{y_obs}{Vector of observed values}

\item{fitted_vals}{Vector of fitted values}

\item{model_matrices}{List of model matrices}
}
\value{
A list with sampled data
}
\description{
Sample model data for large datasets
}
\keyword{internal}
