% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwreg.R
\name{.simulate_from_distribution}
\alias{.simulate_from_distribution}
\title{Simulate observations from a specified distribution family}
\usage{
.simulate_from_distribution(
  n,
  alphaVec,
  betaVec,
  gammaVec,
  deltaVec,
  lambdaVec,
  family
)
}
\arguments{
\item{n}{Number of observations to simulate}

\item{alphaVec}{Vector of alpha parameters}

\item{betaVec}{Vector of beta parameters}

\item{gammaVec}{Vector of gamma parameters}

\item{deltaVec}{Vector of delta parameters}

\item{lambdaVec}{Vector of lambda parameters}

\item{family}{Character string specifying distribution family}
}
\value{
A vector of simulated observations
}
\description{
Simulate observations from a specified distribution family
}
\keyword{internal}
