% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwgof.R
\name{.simulate_p_values_bootstrap}
\alias{.simulate_p_values_bootstrap}
\title{Simulate P-Values Using Parametric Bootstrap}
\usage{
.simulate_p_values_bootstrap(object, n_bootstrap, observed_tests)
}
\arguments{
\item{object}{Object of class "gkwfit"}

\item{n_bootstrap}{Number of bootstrap replicates}

\item{observed_tests}{List of observed test statistics}
}
\value{
A list containing simulated p-values and bootstrap distributions
}
\description{
Simulate P-Values Using Parametric Bootstrap
}
\keyword{internal}
