% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwfit.R
\name{.validate_data}
\alias{.validate_data}
\title{Validate data for GKw family distributions}
\usage{
.validate_data(data, n_params)
}
\arguments{
\item{data}{Numeric vector with values in the (0, 1) interval.}

\item{n_params}{Number of parameters to estimate.}
}
\value{
Validated and possibly adjusted data vector.
}
\description{
Validate data for GKw family distributions
}
\keyword{internal}
