% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwfit.R
\name{.validate_parameters}
\alias{.validate_parameters}
\title{Validate parameters for GKw family distributions}
\usage{
.validate_parameters(start, fixed, param_names)
}
\arguments{
\item{start}{List with initial parameter values.}

\item{fixed}{List of parameters to be held fixed (not estimated).}

\item{param_names}{Character vector of parameter names needed for the model.}
}
\value{
List containing validated start and fixed parameter lists.
}
\description{
Validate parameters for GKw family distributions
}
\keyword{internal}
