% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwfitall.R
\name{get_cdf_function}
\alias{get_cdf_function}
\title{Get the CDF function for a fitted GKw distribution model}
\usage{
get_cdf_function(fit)
}
\arguments{
\item{fit}{An object of class \code{"gkwfit"} from the gkwfit function}
}
\value{
A function that computes CDF values for the fitted distribution
}
\description{
Get the CDF function for a fitted GKw distribution model
}
\keyword{internal}
