% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwfitall.R
\name{get_density_function}
\alias{get_density_function}
\title{Get the appropriate density function for a fitted model}
\usage{
get_density_function(fit)
}
\arguments{
\item{fit}{A gkwfit object}
}
\value{
A function that calculates density values
}
\description{
Get the appropriate density function for a fitted model
}
\keyword{internal}
