% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwfitall.R
\name{get_quantile_function}
\alias{get_quantile_function}
\title{Get the appropriate quantile function for a fitted model}
\usage{
get_quantile_function(fit)
}
\arguments{
\item{fit}{A gkwfit object}
}
\value{
A function that calculates quantile values
}
\description{
Get the appropriate quantile function for a fitted model
}
\keyword{internal}
