% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tmb.R
\name{get_tmb_info}
\alias{get_tmb_info}
\title{Get TMB model information}
\usage{
get_tmb_info(dll_name)
}
\arguments{
\item{dll_name}{The name of the TMB model}
}
\value{
A list with information about the compiled TMB model
}
\description{
Get TMB model information
}
\keyword{internal}
