% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tmb.R
\name{get_tmb_path}
\alias{get_tmb_path}
\title{Get TMB model path}
\usage{
get_tmb_path(dll_name)
}
\arguments{
\item{dll_name}{The name of the TMB model}
}
\value{
The path to the compiled TMB model
}
\description{
Get TMB model path
}
\keyword{internal}
