% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gkwgetstartvalues}
\alias{gkwgetstartvalues}
\title{Main function to estimate GKw distribution parameters using the method of moments.
This implementation is optimized for numerical stability and computational efficiency.}
\usage{
gkwgetstartvalues(x, n_starts = 5L)
}
\arguments{
\item{x}{Data vector (must be in (0,1))}

\item{n_starts}{Number of starting points for optimization}
}
\value{
Vector of estimated parameters \eqn{\alpha, \beta, \gamma, \delta, \lambda}
}
\description{
Main function to estimate GKw distribution parameters using the method of moments.
This implementation is optimized for numerical stability and computational efficiency.
}
