% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwreg.R
\name{list_bounded_datasets}
\alias{list_bounded_datasets}
\title{List all available datasets for bounded response regression}
\usage{
list_bounded_datasets(package = NULL)
}
\arguments{
\item{package}{A character string. If specified, list only datasets from the given package.
Must be one of "betareg" or "simplexreg". If NULL (default), lists datasets from both packages.}
}
\value{
A data frame with names and descriptions of available datasets
}
\description{
List all available datasets for bounded response regression
}
\examples{
\dontrun{
# List all available datasets
list_bounded_datasets()

# List only betareg datasets
list_bounded_datasets("betareg")
}

}
\seealso{
\code{\link{get_bounded_datasets}}
}
