% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwreg-compile-tmb.R
\name{list_tmb_models}
\alias{list_tmb_models}
\title{List Available TMB Models}
\usage{
list_tmb_models(pkg_name = "gkwreg")
}
\arguments{
\item{pkg_name}{Package name (default: "gkwreg")}
}
\value{
Character vector of model names (without .cpp extension)
}
\description{
Returns vector of TMB model names available in the package.
}
