% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pekw}
\alias{pekw}
\title{CDF for Exponentiated Kumaraswamy Distribution}
\usage{
pekw(q, alpha, beta, lambda, lower_tail = TRUE, log_p = FALSE)
}
\arguments{
\item{q}{Vector of quantiles in (0,1).}

\item{alpha}{Shape parameter \eqn{\alpha > 0}.}

\item{beta}{Shape parameter \eqn{\beta > 0}.}

\item{lambda}{Shape parameter \eqn{\lambda > 0}.}

\item{lower_tail}{Logical; if TRUE, returns F(q); else 1-F(q).}

\item{log_p}{Logical; if TRUE, returns log(F(q)) or log(1-F(q)).}
}
\value{
A vector of probabilities of the same length as the broadcast of inputs.
}
\description{
Computes F(x)= P(X ≤ x) for EKw(α, β, λ).
}
\details{
The CDF is
\deqn{
  F(x)= [1 - (1 - x^α)^β ]^λ, \quad 0<x<1.
}
}
