% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pkw}
\alias{pkw}
\title{CDF of the Kumaraswamy Distribution}
\usage{
pkw(q, alpha, beta, lower_tail = TRUE, log_p = FALSE)
}
\arguments{
\item{q}{Vector of quantiles in (0,1).}

\item{alpha}{Shape parameter \eqn{\alpha>0}.}

\item{beta}{Shape parameter \eqn{\beta>0}.}

\item{lower_tail}{If TRUE (default), returns \eqn{F(q)=P(X≤q)}. If FALSE, 1-F(q).}

\item{log_p}{If TRUE, returns log probabilities.}
}
\value{
Vector of probabilities or log-probabilities, matching broadcast of q and parameters.
}
\description{
Computes \eqn{F(q)= P(X ≤ q)} for Kw(\eqn{\alpha,\beta}).
}
\details{
\eqn{F(x)= 1 - (1 - x^\alpha)^\beta}, for 0<x<1.
}
