% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwfitall.R
\name{plot.gkwfitall}
\alias{plot.gkwfitall}
\title{Plot method for gkwfitall objects}
\usage{
\method{plot}{gkwfitall}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"gkwfitall"}}

\item{...}{Additional arguments (currently ignored)}
}
\value{
Invisibly returns the input object
}
\description{
Plot method for gkwfitall objects
}
