% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwgof.R
\name{plot.gkwgof}
\alias{plot.gkwgof}
\title{Plot Method for gkwgof Objects}
\usage{
\method{plot}{gkwgof}(x, title = NULL, ncols = 4, which = NULL, ...)
}
\arguments{
\item{x}{An object of class "gkwgof", typically the result of a call to \code{gkwgof}.}

\item{title}{Plot title}

\item{ncols}{Number of columns to draw plots in graphics window}

\item{which}{A numeric vector specifying which plots to display.
If NULL (default), all available plots will be displayed.}

\item{...}{Additional arguments to be passed to plotting functions.}
}
\value{
The input object \code{x} is returned invisibly.
}
\description{
Creates a panel of diagnostic plots for assessing the goodness-of-fit of a model
from the GKw family of distributions.
}
