% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwfit.R
\name{predict.gkwfit}
\alias{predict.gkwfit}
\title{Calculate predicted density values for a gkwfit object}
\usage{
\method{predict}{gkwfit}(object, newdata = NULL, type = c("density", "cdf"), ...)
}
\arguments{
\item{object}{A gkwfit object}

\item{newdata}{Optional new data points}

\item{type}{Type of prediction: "density" or "cdf"}

\item{...}{Additional arguments}
}
\description{
Calculate predicted density values for a gkwfit object
}
