% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwfit.R
\name{print.anova.gkwfit}
\alias{print.anova.gkwfit}
\title{S3 method for class 'anova.gkwfit'}
\usage{
\method{print}{anova.gkwfit}(
  x,
  digits = max(getOption("digits") - 2L, 3L),
  signif.stars = getOption("show.signif.stars", TRUE),
  ...
)
}
\arguments{
\item{x}{An object of class \code{"anova.gkwfit"}.}

\item{digits}{Minimum number of significant digits to print.}

\item{signif.stars}{Logical; if TRUE, add significance stars.}

\item{...}{Other args passed to}
}
\value{
An object of class \code{"anova.gkwfit"} and prints a summary.
}
\description{
S3 method for class 'anova.gkwfit'
}
