% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwfit.R
\name{print.gkwfit}
\alias{print.gkwfit}
\title{Print method for gkwfit objects}
\usage{
\method{print}{gkwfit}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{A gkwfit object}

\item{digits}{Number of significant digits to display}

\item{...}{Additional arguments passed to print methods}
}
\description{
Print method for gkwfit objects
}
