% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwfitall.R
\name{print.gkwfitall}
\alias{print.gkwfitall}
\title{Print method for gkwfitall objects}
\usage{
\method{print}{gkwfitall}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"gkwfitall"}}

\item{...}{Additional arguments (currently ignored)}
}
\value{
Invisibly returns the input object
}
\description{
Print method for gkwfitall objects
}
\author{
Lopes, J. E.
}
