% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwgof.R
\name{print.gkwgof}
\alias{print.gkwgof}
\title{Print Method for gkwgof Objects}
\usage{
\method{print}{gkwgof}(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{An object of class "gkwgof", typically the result of a call to \code{gkwgof}.}

\item{verbose}{Logical; if TRUE, provides additional details and explanations.
Default is FALSE.}

\item{...}{Additional arguments (currently unused).}
}
\value{
The input object \code{x} is returned invisibly.
}
\description{
Prints a summary of the goodness-of-fit analysis for GKw family distributions.
}
