% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwfit.R
\name{print.summary.gkwfit}
\alias{print.summary.gkwfit}
\title{Print method for summary.gkwfit objects}
\usage{
\method{print}{summary.gkwfit}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{A summary.gkwfit object}

\item{digits}{Number of significant digits to display}

\item{...}{Additional arguments}
}
\description{
Print method for summary.gkwfit objects
}
