% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwfitall.R
\name{print.summary.gkwfitall}
\alias{print.summary.gkwfitall}
\title{Print method for summary.gkwfitall objects}
\usage{
\method{print}{summary.gkwfitall}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"summary.gkwfitall"}}

\item{...}{Additional arguments (currently ignored)}
}
\value{
Invisibly returns the input object
}
\description{
Print method for summary.gkwfitall objects
}
\author{
Lopes, J. E.
}
