% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwgof.R
\name{print.summary.gkwgof}
\alias{print.summary.gkwgof}
\title{Print Method for summary.gkwgof Objects}
\usage{
\method{print}{summary.gkwgof}(x, ...)
}
\arguments{
\item{x}{An object of class "summary.gkwgof".}

\item{...}{Additional arguments (not used).}
}
\value{
The input object invisibly.
}
\description{
Prints a formatted summary of goodness-of-fit results.
}
