% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{qkw}
\alias{qkw}
\title{Quantile Function of the Kumaraswamy Distribution}
\usage{
qkw(p, alpha, beta, lower_tail = TRUE, log_p = FALSE)
}
\arguments{
\item{p}{Vector of probabilities in \link{0,1} (or log scale if log_p=TRUE).}

\item{alpha, }{beta Shape parameters >0.}

\item{lower_tail}{If TRUE, p=F(x); if FALSE, p=1-F(x).}

\item{log_p}{If TRUE, p is log(p).}
}
\description{
For p in (0,1), returns x=Q(p) such that F(x)=p, where F is Kw(\eqn{\alpha,\beta}).
}
\details{
\eqn{Q(p)= {1 - [1 - p]^(1/\beta)}^(1/\alpha)}
}
