% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rkkw}
\alias{rkkw}
\title{Random Generation from KwKw Distribution}
\usage{
rkkw(n, alpha, beta, delta, lambda)
}
\arguments{
\item{n}{Integer number of samples.}

\item{alpha, }{beta, delta, lambda Distribution parameters.}
}
\description{
Generates \code{n} samples from the KwKw(\eqn{\alpha, \beta, \delta, \lambda}) distribution.
}
\details{
The table suggests: if \eqn{V ~ Unif(0,1)}, define
\eqn{U = 1 - (1 - V)^{1/(\delta+1)}}, then
\eqn{X = \{1 - [1 - U^{1/\lambda}]^{1/\beta}\}^{1/\alpha}}.
}
