% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rkw}
\alias{rkw}
\title{Random Generation from the Kumaraswamy Distribution}
\usage{
rkw(n, alpha, beta)
}
\arguments{
\item{n}{Integer number of observations.}

\item{alpha}{Shape parameter \eqn{\alpha>0}.}

\item{beta}{Shape parameter \eqn{\beta>0}.}
}
\value{
A vector of length n with samples in (0,1).
}
\description{
Generates n samples from Kw(\eqn{\alpha,\beta}), using the transformation from Uniform(0,1).
}
\details{
If V ~ Uniform(0,1), then
\eqn{X= \{1 - [1 - V]^{1/\beta}\}^{1/\alpha}}.
}
