% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwfitall.R
\name{summary.gkwfitall}
\alias{summary.gkwfitall}
\title{Summary method for gkwfitall objects}
\usage{
\method{summary}{gkwfitall}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"gkwfitall"}}

\item{...}{Additional arguments (currently ignored)}
}
\value{
A summarized version of the gkwfitall object
}
\description{
Summary method for gkwfitall objects
}
\author{
Lopes, J. E.
}
