% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwgof.R
\name{summary.gkwgof}
\alias{summary.gkwgof}
\title{Summary Method for gkwgof Objects}
\usage{
\method{summary}{gkwgof}(object, ...)
}
\arguments{
\item{object}{An object of class "gkwgof".}

\item{...}{Additional arguments (not used).}
}
\value{
A list containing key summary statistics.
}
\description{
Provides a concise summary of the goodness-of-fit analysis results.
}
