% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwfit.R
\name{BIC.gkwfit}
\alias{BIC.gkwfit}
\title{Calculate BIC for a gkwfit object}
\usage{
\method{BIC}{gkwfit}(object, ...)
}
\arguments{
\item{object}{A gkwfit object}

\item{...}{Additional models}
}
\description{
Calculate BIC for a gkwfit object
}
