% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculateModifiedDevianceResiduals}
\alias{calculateModifiedDevianceResiduals}
\title{Calculate Modified Deviance Residuals}
\usage{
calculateModifiedDevianceResiduals(
  y,
  fitted,
  alpha,
  beta,
  gamma,
  delta,
  lambda
)
}
\arguments{
\item{y}{NumericVector of observations.}

\item{fitted}{NumericVector of fitted values (means).}

\item{alpha}{NumericVector of alpha parameters.}

\item{beta}{NumericVector of beta parameters.}

\item{gamma}{NumericVector of gamma parameters.}

\item{delta}{NumericVector of delta parameters.}

\item{lambda}{NumericVector of lambda parameters.}
}
\value{
NumericVector of modified deviance residuals.
}
\description{
Adjusts deviance residuals to have a distribution closer to N(0,1) by standardizing them.
}
