% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculateQuantileResiduals}
\alias{calculateQuantileResiduals}
\title{Calculate Quantile Residuals}
\usage{
calculateQuantileResiduals(y, alpha, beta, gamma, delta, lambda)
}
\arguments{
\item{y}{NumericVector of observations.}

\item{alpha}{NumericVector of alpha parameters.}

\item{beta}{NumericVector of beta parameters.}

\item{gamma}{NumericVector of gamma parameters.}

\item{delta}{NumericVector of delta parameters.}

\item{lambda}{NumericVector of lambda parameters.}
}
\value{
NumericVector of quantile residuals.
}
\description{
Computes quantile residuals by transforming the cumulative distribution function (CDF) values to the standard normal quantiles.
}
