% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculateScoreResiduals}
\alias{calculateScoreResiduals}
\title{Calculate Score Residuals}
\usage{
calculateScoreResiduals(y, fitted, alpha, beta, gamma, delta, lambda)
}
\arguments{
\item{y}{NumericVector of observations.}

\item{fitted}{NumericVector of fitted values (means).}

\item{alpha}{NumericVector of alpha parameters.}

\item{beta}{NumericVector of beta parameters.}

\item{gamma}{NumericVector of gamma parameters.}

\item{delta}{NumericVector of delta parameters.}

\item{lambda}{NumericVector of lambda parameters.}
}
\value{
NumericVector of score residuals.
}
\description{
Computes score residuals based on the numerical derivative (score) of the log-likelihood with respect to the observation.
}
