% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwfit.R
\name{coef.gkwfit}
\alias{coef.gkwfit}
\title{Extract coefficients from a gkwfit object}
\usage{
\method{coef}{gkwfit}(object, ...)
}
\arguments{
\item{object}{A gkwfit object}

\item{...}{Additional arguments}
}
\description{
Extract coefficients from a gkwfit object
}
