% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwfit.R
\name{confint.gkwfit}
\alias{confint.gkwfit}
\title{Extract confidence intervals from a gkwfit object}
\usage{
\method{confint}{gkwfit}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{A gkwfit object}

\item{parm}{Parameters to extract confidence intervals for}

\item{level}{Confidence level}

\item{...}{Additional arguments}
}
\description{
Extract confidence intervals from a gkwfit object
}
