% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwfitall.R
\name{create_comparison_plots}
\alias{create_comparison_plots}
\title{Create comparison plots of all fitted distributions}
\usage{
create_comparison_plots(fits, data)
}
\arguments{
\item{fits}{List of fitted gkwfit models}

\item{data}{Original data vector}
}
\value{
A ggplot2 object with multiple panels
}
\description{
Create comparison plots of all fitted distributions
}
\keyword{internal}
