% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwfitall.R
\name{create_comparison_table}
\alias{create_comparison_table}
\title{Create comparison table of fit statistics with expanded metrics}
\usage{
create_comparison_table(fits, gof_tests = c("ks", "ad", "cvm"))
}
\arguments{
\item{fits}{List of fitted gkwfit models}

\item{gof_tests}{Character vector of goodness-of-fit tests to perform}
}
\value{
Data frame with comparison statistics
}
\description{
Create comparison table of fit statistics with expanded metrics
}
\keyword{internal}
