% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dbeta_}
\alias{dbeta_}
\title{Density of the Beta Distribution}
\usage{
dbeta_(x, gamma, delta, log_prob = FALSE)
}
\arguments{
\item{x}{Vector of quantiles in (0,1).}

\item{gamma}{First shape parameter > 0.}

\item{delta}{Second shape parameter > 0.}

\item{log_prob}{If TRUE, returns log-density.}
}
\value{
A vector of densities (or log-densities).
}
\description{
Computes the PDF of Beta(\code{gamma, delta}) for x in (0,1).
}
\details{
The PDF is
\deqn{
  f(x) = \frac{x^{\gamma-1}\,(1-x)^{\delta}}
             {B(\gamma, \delta+1)}, \quad 0<x<1.
}
}
