% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dkw}
\alias{dkw}
\title{PDF of the Kumaraswamy Distribution}
\usage{
dkw(x, alpha, beta, log_prob = FALSE)
}
\arguments{
\item{x}{Vector of quantiles in (0,1).}

\item{alpha}{Shape parameter \eqn{\alpha>0}.}

\item{beta}{Shape parameter \eqn{\beta>0}.}

\item{log_prob}{If TRUE, returns log of the PDF.}
}
\value{
A vector of (log-)densities, same length as broadcast of x and parameters.
}
\description{
Computes the PDF of Kw(\eqn{\alpha,\beta}) at values x in (0,1).
}
\details{
PDF: \eqn{f(x)= \alpha\,\beta \; x^{\alpha-1}\,(1 - x^\alpha)^{\beta-1}}, for 0<x<1.
}
