% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dmc}
\alias{dmc}
\title{Density of the Beta Power Distribution}
\usage{
dmc(x, gamma, delta, lambda, log_prob = FALSE)
}
\arguments{
\item{x}{Vector of quantiles in (0,1).}

\item{gamma}{Shape parameter \eqn{γ > 0}.}

\item{delta}{Shape parameter \eqn{δ \ge 0}.}

\item{lambda}{Shape parameter \eqn{λ > 0}.}

\item{log_prob}{Logical; if TRUE, returns log-density.}
}
\value{
A vector of densities or log-densities, the same length as the broadcast of inputs.
}
\description{
Computes the PDF of the Beta Power (BP or MC) distribution with parameters \eqn{γ, δ, λ}.
}
\details{
The PDF is
\deqn{
  f(x) = \frac{\lambda}{B(\gamma,\delta+1)} \; x^{\gamma \lambda - 1} \; (1 - x^\lambda)^\delta, \quad 0<x<1.
}
with \eqn{B(\gamma,\delta+1)} the Beta function. This sub-family arises
from the Generalized Kumaraswamy (GKw) by setting \eqn{\alpha=1,\beta=1}.
}
