% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwreg-utils.R
\name{.get_family_param_info}
\alias{.get_family_param_info}
\title{Get Parameter Information for a GKw Family Distribution}
\usage{
.get_family_param_info(family)
}
\arguments{
\item{family}{The GKw family distribution name.}
}
\value{
A list with parameter information.
}
\description{
Get Parameter Information for a GKw Family Distribution
}
\keyword{internal}
