% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwfit.R, R/gkwreg.R
\name{.get_family_param_info}
\alias{.get_family_param_info}
\title{Get family parameter information}
\usage{
.get_family_param_info(family)

.get_family_param_info(family)
}
\arguments{
\item{family}{The GKw family distribution name.}
}
\value{
List with parameter names and count.

A list with parameter information.
}
\description{
Get family parameter information

Get Parameter Information for a GKw Family Distribution
}
\keyword{internal}
