% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwreg-utils.R
\name{.validate_data}
\alias{.validate_data}
\title{Validate Data for GKw Regression}
\usage{
.validate_data(data, n_params)
}
\arguments{
\item{data}{Numeric vector to validate.}

\item{n_params}{Number of parameters in the selected model.}
}
\value{
The validated data.
}
\description{
Validate Data for GKw Regression
}
\keyword{internal}
