% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwfitall.R
\name{get_cdf_function}
\alias{get_cdf_function}
\title{Get the appropriate CDF function for a fitted model}
\usage{
get_cdf_function(fit)
}
\arguments{
\item{fit}{A gkwfit object}
}
\value{
A function that calculates CDF values
}
\description{
Get the appropriate CDF function for a fitted model
}
\keyword{internal}
