% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwfitall.R
\name{get_density_function}
\alias{get_density_function}
\title{Get the density function for a fitted GKw distribution model}
\usage{
get_density_function(fit)
}
\arguments{
\item{fit}{An object of class \code{"gkwfit"} from the gkwfit function}
}
\value{
A function that computes density values for the fitted distribution
}
\description{
Get the density function for a fitted GKw distribution model
}
\keyword{internal}
