% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tmb.R
\name{list_tmb_models}
\alias{list_tmb_models}
\title{List all available TMB model files}
\usage{
list_tmb_models(pkg_name = "gkwreg")
}
\arguments{
\item{pkg_name}{The package name (default: "gkwreg")}
}
\value{
A character vector with the names of available TMB model files (without extension)
}
\description{
List all available TMB model files
}
