% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwfit.R
\name{logLik.gkwfit}
\alias{logLik.gkwfit}
\title{Extract log-likelihood from a gkwfit object}
\usage{
\method{logLik}{gkwfit}(object, ...)
}
\arguments{
\item{object}{A gkwfit object}

\item{...}{Additional arguments}
}
\description{
Extract log-likelihood from a gkwfit object
}
