% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pkkw}
\alias{pkkw}
\title{CDF of the KwKw Distribution}
\usage{
pkkw(q, alpha, beta, delta, lambda, lower_tail = TRUE, log_p = FALSE)
}
\arguments{
\item{q}{Vector of quantiles in \eqn{(0,1)}.}

\item{alpha, }{beta, delta, lambda Parameters (see \code{\link{dkkw}}).}

\item{lower_tail}{Logical; if TRUE (default), returns F(q)=P(X≤q), else 1-F(q).}

\item{log_p}{Logical; if TRUE, returns log of the probability.}
}
\description{
Computes the CDF for KwKw(\eqn{\alpha, \beta, \delta, \lambda}), i.e. P(X≤x).
}
\details{
The CDF is
\deqn{
  F(x) = 1 - \bigl\{1 - \bigl[1 - (1 - x^\alpha)^\beta\bigr]^\lambda\bigr\}^{\delta + 1}.
}
}
