% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwfit.R
\name{plot.gkwfit}
\alias{plot.gkwfit}
\title{Plot method for gkwfit objects}
\usage{
\method{plot}{gkwfit}(
  x,
  which = c("histogram", "pp", "qq", "profile"),
  ask = (which != "all" && length(which) > 1 && dev.interactive()),
  ...
)
}
\arguments{
\item{x}{A gkwfit object}

\item{which}{Which plots to display}

\item{ask}{Whether to ask before displaying each plot}

\item{...}{Additional arguments}
}
\description{
Plot method for gkwfit objects
}
