% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{qekw}
\alias{qekw}
\title{Quantile Function for Exponentiated Kumaraswamy}
\usage{
qekw(p, alpha, beta, lambda, lower_tail = TRUE, log_p = FALSE)
}
\arguments{
\item{p}{Vector of probabilities in \link{0,1}.}

\item{alpha, }{beta, lambda Distribution parameters.}

\item{lower_tail}{Logical; if TRUE, p is F(x). If FALSE, p=1-F(x).}

\item{log_p}{Logical; if TRUE, p is given as log(p).}
}
\description{
Computes the quantiles x = Q(p) for EKw(α, β, λ).
}
\details{
The inverse of F(x)= \link{1 - (1 - x^α)^β }^λ is
\deqn{
  Q(p)= \Bigl\{1 - \bigl[1 - p^{1/\lambda}\bigr]^{1/\beta}\Bigr\}^{1/\alpha}.
}
if lower_tail=TRUE. Adjust accordingly if \code{lower_tail=FALSE} or \code{log_p=TRUE}.
}
