% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{qkkw}
\alias{qkkw}
\title{Quantile Function of the KwKw Distribution}
\usage{
qkkw(p, alpha, beta, delta, lambda, lower_tail = TRUE, log_p = FALSE)
}
\arguments{
\item{p}{Vector of probabilities in (0,1).}

\item{alpha, }{beta, delta, lambda Distribution parameters (see \code{\link{dkkw}}).}

\item{lower_tail}{Logical; if TRUE, p is F(x). If FALSE, p is 1-F(x).}

\item{log_p}{Logical; if TRUE, p is given as log(p).}
}
\description{
Computes the quantiles for KwKw(\eqn{\alpha, \beta, \delta, \lambda}).
}
\details{
Invert the CDF
\eqn{
  F(x)= 1 - [1 - (1 - x^α)^β]^λ)^(δ+1).
}
The resulting formula is
\deqn{
  Q(p)= \Bigl\{1 - \Bigl[1 - \Bigl(1 - p\Bigr)^{\frac{1}{\delta+1}}\Bigr]^{\frac{1}{\lambda}}\Bigr\}^{\frac{1}{\beta}}\Bigr\}^{\frac{1}{\alpha}}.
}
}
