% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{qmc}
\alias{qmc}
\title{Quantile Function of the Beta Power Distribution}
\usage{
qmc(p, gamma, delta, lambda, lower_tail = TRUE, log_p = FALSE)
}
\arguments{
\item{p}{Vector of probabilities in (0,1) (or in log scale if log_p=TRUE).}

\item{gamma}{Shape parameter \eqn{γ > 0}.}

\item{delta}{Shape parameter \eqn{δ \ge 0}.}

\item{lambda}{Shape parameter \eqn{λ > 0}.}

\item{lower_tail}{Logical; if TRUE, p=F(x). If FALSE, p=1-F(x).}

\item{log_p}{Logical; if TRUE, p is log(p).}
}
\description{
Computes \eqn{x = Q(p)}, the quantile function for \eqn{BP(γ, δ, λ)}.
}
\details{
The quantile function is
\deqn{
  Q(p) = \Bigl[\mathrm{qbeta}(p,\;\gamma,\;\delta+1)\Bigr]^{1/\lambda}.
}
if lower_tail=TRUE and log_p=FALSE. Adjust for tails or logs accordingly.
}
