% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rekw}
\alias{rekw}
\title{Random Generation for Exponentiated Kumaraswamy Distribution}
\usage{
rekw(n, alpha, beta, lambda)
}
\arguments{
\item{n}{Integer number of observations.}

\item{alpha}{Shape parameter \eqn{\alpha > 0}.}

\item{beta}{Shape parameter \eqn{\beta > 0}.}

\item{lambda}{Shape parameter \eqn{\lambda > 0}.}
}
\description{
Generates n samples from EKw(α, β, λ).
}
\details{
Implementation via the quantile method:  X = Q(U),  U ~ Uniform(0,1).
The quantile is
\deqn{
  Q(u)= \Bigl\{1 - \bigl[1 - u^{1/\lambda}\bigr]^{1/\beta}\Bigr\}^{1/\alpha}.
}
}
