% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmc}
\alias{rmc}
\title{Random Generation from Beta Power Distribution}
\usage{
rmc(n, gamma, delta, lambda)
}
\arguments{
\item{n}{Integer number of observations.}

\item{gamma}{Shape parameter \eqn{γ > 0}.}

\item{delta}{Shape parameter \eqn{δ \ge 0}.}

\item{lambda}{Shape parameter \eqn{λ > 0}.}
}
\value{
A numeric vector of length n with random draws in (0,1).
}
\description{
Generates n samples from BP(γ, δ, λ) by drawing U ~ Beta(γ, δ+1) then X= U^(1/λ).
}
\details{
The distribution arises from GKw with \eqn{\alpha=1,\beta=1}. This is sometimes
also called the Beta-Power distribution in the literature.
}
