% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwfit.R
\name{summary.gkwfit}
\alias{summary.gkwfit}
\title{Summary method for gkwfit objects}
\usage{
\method{summary}{gkwfit}(object, ...)
}
\arguments{
\item{object}{A gkwfit object}

\item{...}{Additional arguments}
}
\description{
Summary method for gkwfit objects
}
