% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkwfit.R
\name{vcov.gkwfit}
\alias{vcov.gkwfit}
\title{Extract variance-covariance matrix from a gkwfit object}
\usage{
\method{vcov}{gkwfit}(object, ...)
}
\arguments{
\item{object}{A gkwfit object}

\item{...}{Additional arguments}
}
\description{
Extract variance-covariance matrix from a gkwfit object
}
